added

+++ Modules/Case/Database/Migrations/2022_08_11_121300_add_released_date_into_case_model_table.php
@@ -14,8 +14,8 @@ class AddReleasedDateIntoCaseModelTable extends Migration
     public function up()
     {
         Schema::table('case_models', function (Blueprint $table) {
-            $table->tinyInteger('released_year')->nullable()->after('status');
-            $table->tinyInteger('released_quarter')->nullable()->after('released_year');
+            $table->unsignedInteger('released_year')->nullable()->after('status');
+            $table->unsignedInteger('released_quarter')->nullable()->after('released_year');
         });
     }
